### 2026-02-16 #10042 @editorialbot
[PRE REVIEW]: PointCloudCrafter: Tools for Handling, Manipulating and Analyzing of Point Clouds
**Submitting author:** @mleitenstern (<a href="http://orcid.org/0009-0008-6436-7967">Maximilian Leitenstern</a>)
**Repository:** https://github.com/TUMFTM/PointCloudCrafter
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.4
