### 2026-02-17 7fcdc38
remove Git LFS (#56)
