+++ tests/rosbag/metadata.yaml
@@ -1,45 +1,0 @@
-rosbag2_bagfile_information:
-  version: 5
-  storage_identifier: ""
-  duration:
-    nanoseconds: 149900156
-  starting_time:
-    nanoseconds_since_epoch: 1756285989513751040
-  message_count: 13
-  topics_with_message_count:
-    - topic_metadata:
-        name: /edgar/sensor/lidar/innovusion/frontcenter/points
-        type: sensor_msgs/msg/PointCloud2
-        serialization_format: cdr
-        offered_qos_profiles: "- history: 1\n  depth: 1\n  reliability: 2\n  durability: 2\n  deadline:\n    sec: 9223372036\n    nsec: 854775807\n  lifespan:\n    sec: 9223372036\n    nsec: 854775807\n  liveliness: 1\n  liveliness_lease_duration:\n    sec: 9223372036\n    nsec: 854775807\n  avoid_ros_namespace_conventions: false"
-      message_count: 3
-    - topic_metadata:
-        name: /edgar/sensor/lidar/innovusion/rearcenter/points
-        type: sensor_msgs/msg/PointCloud2
-        serialization_format: cdr
-        offered_qos_profiles: "- history: 1\n  depth: 1\n  reliability: 2\n  durability: 2\n  deadline:\n    sec: 9223372036\n    nsec: 854775807\n  lifespan:\n    sec: 9223372036\n    nsec: 854775807\n  liveliness: 1\n  liveliness_lease_duration:\n    sec: 9223372036\n    nsec: 854775807\n  avoid_ros_namespace_conventions: false"
-      message_count: 3
-    - topic_metadata:
-        name: /edgar/sensor/lidar/ouster/frontleft/points
-        type: sensor_msgs/msg/PointCloud2
-        serialization_format: cdr
-        offered_qos_profiles: "- history: 1\n  depth: 5\n  reliability: 2\n  durability: 2\n  deadline:\n    sec: 9223372036\n    nsec: 854775807\n  lifespan:\n    sec: 9223372036\n    nsec: 854775807\n  liveliness: 1\n  liveliness_lease_duration:\n    sec: 9223372036\n    nsec: 854775807\n  avoid_ros_namespace_conventions: false"
-      message_count: 3
-    - topic_metadata:
-        name: /edgar/sensor/lidar/ouster/frontright/points
-        type: sensor_msgs/msg/PointCloud2
-        serialization_format: cdr
-        offered_qos_profiles: "- history: 1\n  depth: 5\n  reliability: 2\n  durability: 2\n  deadline:\n    sec: 9223372036\n    nsec: 854775807\n  lifespan:\n    sec: 9223372036\n    nsec: 854775807\n  liveliness: 1\n  liveliness_lease_duration:\n    sec: 9223372036\n    nsec: 854775807\n  avoid_ros_namespace_conventions: false"
-      message_count: 3
-    - topic_metadata:
-        name: /tf_static
-        type: tf2_msgs/msg/TFMessage
-        serialization_format: cdr
-        offered_qos_profiles: "- history: 1\n  depth: 1\n  reliability: 1\n  durability: 1\n  deadline:\n    sec: 9223372036\n    nsec: 854775807\n  lifespan:\n    sec: 9223372036\n    nsec: 854775807\n  liveliness: 1\n  liveliness_lease_duration:\n    sec: 9223372036\n    nsec: 854775807\n  avoid_ros_namespace_conventions: false"
-      message_count: 1
-  compression_format: ""
-  compression_mode: ""
-  relative_file_paths:
-    - rosbag_0.mcap
-  files:
-    []
