+++ docs_cpp/search/variables_5.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['group_5f_0',['group_',['../classCLI_1_1OptionBase.html#ad0d8bd0637023f7d87a8e95315e8fb21',1,'CLI::OptionBase::group_'],['../classCLI_1_1App.html#ac31738f617c3ce7aaf9fd94c3f54fc0f',1,'CLI::App::group_']]]
+];
