+++ docs_cpp/search/functions_c.js
@@ -1,0 +1,22 @@
+var searchData=
+[
+  ['make_5fdescription_0',['make_description',['../classCLI_1_1Formatter.html#ac74cd92603dfd527bf28712f420329cf',1,'CLI::Formatter']]],
+  ['make_5fexpanded_1',['make_expanded',['../classCLI_1_1Formatter.html#aff3030235805482244475b1d706a1306',1,'CLI::Formatter']]],
+  ['make_5ffooter_2',['make_footer',['../classCLI_1_1Formatter.html#a7f7a77cd112d6c279ac1175a8ebd3e6d',1,'CLI::Formatter']]],
+  ['make_5fgroup_3',['make_group',['../classCLI_1_1Formatter.html#a7437125efe9b3516a3313547b31f813e',1,'CLI::Formatter']]],
+  ['make_5fgroups_4',['make_groups',['../classCLI_1_1Formatter.html#a19cff35f1e11d0322cafc9dc7a0cdc6a',1,'CLI::Formatter']]],
+  ['make_5fhelp_5',['make_help',['../classCLI_1_1FormatterBase.html#ac8ab93655a645352634709e15087080b',1,'CLI::FormatterBase::make_help()'],['../classCLI_1_1FormatterLambda.html#a5eca71447c56251c67b97187e4a888e6',1,'CLI::FormatterLambda::make_help()'],['../classCLI_1_1Formatter.html#a114f6b9a0c1e7765cc57b34dff2ad9f2',1,'CLI::Formatter::make_help(const App *, std::string, AppFormatMode) const override']]],
+  ['make_5foption_6',['make_option',['../classCLI_1_1Formatter.html#a1503e498cef1cab66a01d84bac59b5d8',1,'CLI::Formatter']]],
+  ['make_5foption_5fdesc_7',['make_option_desc',['../classCLI_1_1Formatter.html#acb3da9c5d8c3637b915987abd1599734',1,'CLI::Formatter']]],
+  ['make_5foption_5fname_8',['make_option_name',['../classCLI_1_1Formatter.html#a1f9291269b6daf74361222e1149c1afb',1,'CLI::Formatter']]],
+  ['make_5foption_5fopts_9',['make_option_opts',['../classCLI_1_1Formatter.html#aea58f88b1c39121b81540c0685439a93',1,'CLI::Formatter']]],
+  ['make_5foption_5fusage_10',['make_option_usage',['../classCLI_1_1Formatter.html#a81a2b22412c1a9cbf52caee6465fdad2',1,'CLI::Formatter']]],
+  ['make_5fpositionals_11',['make_positionals',['../classCLI_1_1Formatter.html#aa1f17baeee2e2d6ab30297fe5faf8552',1,'CLI::Formatter']]],
+  ['make_5fsubcommand_12',['make_subcommand',['../classCLI_1_1Formatter.html#a9f523ad9792c9c2231c974934c853a5a',1,'CLI::Formatter']]],
+  ['make_5fsubcommands_13',['make_subcommands',['../classCLI_1_1Formatter.html#aaeb7e461f2e1afe437b53d4990cd3422',1,'CLI::Formatter']]],
+  ['make_5fusage_14',['make_usage',['../classCLI_1_1Formatter.html#a1e9b2e813a3b17ba9d50277d557a3a9b',1,'CLI::Formatter']]],
+  ['mandatory_15',['mandatory',['../classCLI_1_1OptionBase.html#a083eb9ce01d6f8a59f6cfea0cece2e05',1,'CLI::OptionBase']]],
[…]
