### 2026-02-17 #10042 @nitishsanghi
Hello Maximilian,

I will review later today.

Thanks,
Nitish

On Tue, 17 Feb 2026 at 01:07, Maximilian Leitenstern <
***@***.***> wrote:

> *mleitenstern* left a comment (openjournals/joss-reviews#10042)
> <https://github.com/openjournals/joss-reviews/issues/10042#issuecomment-3913220314>
>
> Searching the list of reviewers, we found three potential candidates with
> expertise in C++/ROS2/Robotics:
>
>    - @nitishsanghi <https://github.com/nitishsanghi> (Nitish Sanghi)
>    - @mcbed <https://github.com/mcbed> (Maciej Bednarczyk)
>    - @KennethKwabenaKenney <https://github.com/KennethKwabenaKenney>
>    (Kenneth Kwabena Kenney)
>
> We'd appreciate it, if any of you could review our work.
> Thanks in advance!
>
