### 2026-02-18 #10042 @KennethKwabenaKenney
@nitishsanghi
I guess we will need an editor first before review can begin

I don't know if I am right about that, but from my other JOSS review we have an editor overseeing this

However I don't know if it's possible to start reviewing while there's is no editor

@mleitenstern
I also don't know how the system assigns editors, and if it's possible to suggest your own editor and the system will acknowledge that

Or you have to wait until an editor is assigned
