### 2026-02-18 #10042 @mleitenstern
Hi @nitishsanghi and @KennethKwabenaKenney ,

thanks for your fast response and the kindness to review the paper!
As it is our first submission to JOSS, we are not entirely familiar with the review process (at least apart from whats in the docs).

As you can see at the top of this issue, there hasn't been an editor assigned so far and we were asked to suggest reviewers based on the available list (where we found you based on your area of expertise in C++, ROS2, Robotics).

I'd wait a day or two from now on for an editor to be assigned here, otherwise i'll try to reach out ?
