### 2026-03-30 #10042 @crvernon
@mleitenstern please ensure that a [state of the field section](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain) exists within the paper.
