### 2026-03-30 #10042 @mleitenstern
@crvernon done - we included the state of the field section in the statement of need section.
Now it is split.
