### 2026-04-14 86c306b
fix entrypoint
