+++ .github/workflows/pypi.yml
@@ -18,7 +18,7 @@ jobs:
       id-token: write
     strategy:
       matrix:
-        DISTRO: [jammy]
+        DISTRO: [noble]
         TOOLCHAIN: [gcc]
     steps:
       - name: Checkout Repo
