+++ docker/Dockerfile
@@ -70,7 +70,7 @@ RUN /bin/bash -c '. /opt/ros/$ROS_DISTRO/setup.bash && \
 # Set entrypoint by sourcing overlay workspace
 RUN echo 'source "/opt/ros/$ROS_DISTRO/setup.bash"' >> /root/.bashrc && \
     echo '. /ros_ws/install/setup.bash' >> /root/.bashrc && \
-    echo '#!/bin/bash\nset -e\n\n# setup ros environment\nsource "/opt/ros/$ROS_DISTRO/setup.bash"\n. /ros_ws/install/setup.bash\nexec "$@"' > /ros_entrypoint.sh && \
+    printf '#!/bin/bash\nset -e\n\n# setup ros environment\nsource "/opt/ros/$ROS_DISTRO/setup.bash"\n. /ros_ws/install/setup.bash\nexec "$@"\n' > /ros_entrypoint.sh && \
     chmod +x /ros_entrypoint.sh
 
 ENTRYPOINT ["/ros_entrypoint.sh"]
