### 2026-04-14 9efcc56
Revert "fix wheel deps"
This reverts commit cb2920dca82a811d1f787f8471919920f9b1831c.
