### 2026-04-14 07982c7
fix readme and baked libs
