### 2026-04-14 aa55bb3
try supporting python3.10 and ubuntu22 again after fixing pypi deps
