+++ .github/workflows/testpypi.yml
@@ -14,6 +14,9 @@ jobs:
       fail-fast: false
       matrix:
         include:
+          - python_version: "3.10"
+            python_tag: "cp310"
+            ubuntu_test: "22.04"
           - python_version: "3.12"
             python_tag: "cp312"
             ubuntu_test: "24.04"
@@ -71,6 +74,9 @@ jobs:
       fail-fast: false
       matrix:
         include:
+          - python_version: "3.10"
+            python_tag: "cp310"
+            ubuntu_version: "22.04"
           - python_version: "3.12"
             python_tag: "cp312"
             ubuntu_version: "24.04"
