### 2026-04-14 bee0a70
Update docker/Dockerfile
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
