+++ docker/Dockerfile
@@ -30,7 +30,8 @@ RUN apt-get update && apt-get install -y \
 # Add ROS 2 GPG key and repository
 RUN export ROS_APT_SOURCE_VERSION=$(curl -s https://api.github.com/repos/ros-infrastructure/ros-apt-source/releases/latest | grep -F "tag_name" | awk -F\" '{print $4}') && \
   curl -L -o /tmp/ros2-apt-source.deb "https://github.com/ros-infrastructure/ros-apt-source/releases/download/${ROS_APT_SOURCE_VERSION}/ros2-apt-source_${ROS_APT_SOURCE_VERSION}.$(. /etc/os-release && echo ${VERSION_CODENAME})_all.deb" && \
-  dpkg -i /tmp/ros2-apt-source.deb
+  dpkg -i /tmp/ros2-apt-source.deb && \
+  rm -f /tmp/ros2-apt-source.deb
 
 # Install dependencies
 RUN apt-get update && apt-get install -y -q --no-install-recommends \
