+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "scikit_build_core.build"
 
 [project]
 name = "pointcloudcrafter"
-version = "1.2.5"
+version = "1.3.4"
 description = "Tools for Handling, Manipulating and Analyzing of Point Clouds"
 readme = "README.md"
 authors = [
