### 2026-04-23 #10042 @jeofo
> 👋 [@nitishsanghi](https://github.com/nitishsanghi), [@KennethKwabenaKenney](https://github.com/KennethKwabenaKenney), [@jeofo](https://github.com/jeofo), [@mcbed](https://github.com/mcbed), would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html. Please be so kind to let me know even if you're not interested, so that I can start looking for other reviewers.

Yes. I'd be happy to review.
