### 2026-04-23 c71a981
reduce pcds for file testing
