+++ .github/workflows/pypi.yml
@@ -224,6 +224,11 @@ jobs:
               echo 'Test: Split option'
               pointcloudcrafter-rosbag \$TEST_ROSBAG \$OUT_DIR \$TOPIC1 \$TOPIC2 --split 5.0
 
+              # Cleanup
+              echo 'Cleanup'
+              rm -rf \$OUT_DIR/*
+              pointcloudcrafter-rosbag \$TEST_ROSBAG \$OUT_DIR \$TOPIC1 \$TOPIC2 -m 2
+
               # ========================================
               # pointcloudcrafter-file tests
               # ========================================
