### 2026-04-24 2f9b29d
save the splitted clouds into separate directories
