### 2026-04-24 2957831
49 further functions to implement (#61)
* --split option added

* reduce pcds for file testing

* save the splitted clouds into separate directories
