### 2026-04-24 a02dd9c
fix lib linking issue
