+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "scikit_build_core.build"
 
 [project]
 name = "pointcloudcrafter"
-version = "1.4.0"
+version = "1.4.1"
 description = "Tools for Handling, Manipulating and Analyzing of Point Clouds"
 readme = "README.md"
 authors = [
