### 2026-04-24 e1e325e
update docs + set version number
