### 2026-04-28 #10459 @nitishsanghi
@jolars Does the order of the sections in the paper matter? JOSS reviewer criteria and checklist have 'Statement of need' before 'State of the field' but the submitted paper switches the order. Is that acceptable?
