### 2026-04-28 #10459 @nitishsanghi
@jolars This is a minor thing, but the authors mention 'CloudCompare' for comparison but do not include any document in references to confirm claim. Maybe OK without a reference as it is a known tool but since it is a comparison may want to add some reference document.
