### 2026-04-28 #10459 @nitishsanghi
@jolars Evaluating the project for 'Collaborative Effort' is a bit tricky based on JOSS review criteria. This repo seems to have 2 contributors and no external contributions. All issues were also opened/resolved by the authors themselves. The paper does mention some research impact with publications and internal use in the research group. JOSS review criteria mentions that is acceptable. Do we need to request more evidence? I am leaning towards checking off 'Collaborative Effort' based on the paper mentioning internal use in the research group and the couple of publications where it was used. I would suggest the authors to elaborate on internal use for strengthening this criteria.
