### 2026-04-28 #10459 @nitishsanghi
@jolars The authors have followed 'Good practices' mostly, but 'contribution/support pathways' are not clearly specified it seems. Can we request the authors to do so?
