### 2026-04-28 #10459 @jolars
> [@jolars](https://github.com/jolars) The authors have followed 'Good practices' mostly, but 'contribution/support pathways' are not clearly specified it seems. Can we request the authors to do so?

Absolutely. Best way is to raise an issue at the repo and reference it with `openjournals/joss-reviews#10459`.
