### 2026-04-29 TUMFTM/PointCloudCrafter#66 @nitishsanghi
openjournals/joss-reviews#10459 - 'contribution/support pathways' are not clearly specified
There is requirement for clear community contribution guidelines.

[Review Criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines)
