### 2026-04-29 TUMFTM/PointCloudCrafter#67 @nitishsanghi
openjournals/joss-reviews#10459: Missing reference for cloudcompare
Submitted paper references cloudcompare for comparison. No reference documentation added to support claim.
