### 2026-04-29 TUMFTM/PointCloudCrafter#68 @nitishsanghi
openjournals/joss-reviews#10459: Collaborative Effort criteria requires more evidence
Paper mentions a couple of papers published and states the repo/library has been used internally for projects. Please provide more details on what those projects were or how the research group/students used or applied this library. Collaborative effort is an important criteria and requires sufficient evidence for review.
