### 2026-04-29 #10459 @nitishsanghi
Issues created:
[openjournals/joss-reviews#10459 - 'contribution/support pathways' are not clearly specified #66](https://github.com/TUMFTM/PointCloudCrafter/issues/66)
[openjournals/joss-reviews#10459: Missing reference for cloudcompare #67](https://github.com/TUMFTM/PointCloudCrafter/issues/67)
[openjournals/joss-reviews#10459: Collaborative Effort criteria requires more evidence #68](https://github.com/TUMFTM/PointCloudCrafter/issues/68)
