### 2026-05-06 1cbdd02
Add option to save to rosbag
