### 2026-05-06 def7fcc
Add option to save to rosbag
