### 2026-05-07 #10459 @jeofo
@mleitenstern  Thanks for the submission! Just a few things and formalities after an initial look.

While setting up and testing, I ran into a few issues:

1. On macOS 26.3.1 (arm64), running the paper's documented install command `pip install pointcloudcrafter` in a clean Python venv fails immediately which is expected since PyPI currently only hosts only Linux x86_64 wheels for the package and there's no macOS wheel, no aarch64 wheel, no Windows wheel, and no source distribution for fall back. Specifically, the latest release (1.4.1) ships exactly one file: pointcloudcrafter-1.4.1-cp312-cp312-manylinux_2_39_x86_64.whl. However, this is a bit unclear by just looking at the project since:
- The README badge "Tested on Ubuntu 22.04 | 24.04" and the paper's Availability section give the impression of broader Linux/Unix support. Even the listed Ubuntu 22.04 ships glibc 2.35, so a manylinux_2_39 wheel cannot install there (untested, but it shouldn't)
- pyproject.toml declares requires-python = ">=3.10", but 1.4.1 ships only a cp312 wheel, so users on Python 3.10 / 3.11 / 3.13 / 3.14 on otherwise-supported Linux also fail to install.

Since the project's own classifier Operating System :: POSIX :: Linux in pyproject.toml already concedes Linux-only, I think we need to tighten the claims to match what is shipped. Maybe drop the "Ubuntu 22.04" portion of the README badge, set requires-python = ">=3.12", and update the paper Availability section.

2. Following the README  "Test Data" section verbatim/strictly, `mv /tmp/rosbag-test/tests /path/to/pointcloudcrafter-root/.` fails because the repo already ships a non-empty tests/ directory. Switching the instruction to either rsync or polishing this line would make things a bit less confusing.

3. The "Getting started" docs page (docs/getting_started.md) still tells users to run GIT_LFS_SKIP_SMUDGE=1 git clone … and git lfs pull, but Git LFS was removed in a closed and merged issue 56 and .gitattributes is empty (0 bytes), so those steps don't actually do anything and might confuse new users.

Some other changes I suggest making for a more complete submission:

- The paper currently runs ~1600 words, while the JOSS rubric targets 250–1000.
- Version metadata is out of sync: pyproject.toml is at 1.4.2, but package.xml is still at 1.2.0, while the latest GitHub release / PyPI version is 1.4.1. Worth aligning all three on 1.4.1.
- In the State-of-the-field section, CloudCompare is named alongside PCL/Open3D/PDAL/etc. but is the only one without a citation; adding a entry would make the comparison consistent.
- There's no `CONTRIBUTING.md` or `CODE_OF_CONDUCT.md` in the repo. Per the JOSS rubric on community guidelines, contributors and bug-reporters should have an explicit pathway.

Once these are addressed, I think the submission will be in better shape. I'll try to setup your project on a Linux box and will let you know if I find more suggestions.

Nice work on the multi-distribution story (PyPI / Docker / ROS2 / source) and the Doxygen + MkDocs setup. Looks like a super cool project rather than just a JOSS submission.
