+++ CODE_OF_CONDUCT.md
@@ -1,0 +1,86 @@
+
+# Contributor Covenant Code of Conduct
+
+## Our Pledge
+
+We pledge to make our community welcoming, safe, and equitable for all.
+
+We are committed to fostering an environment that respects and promotes the dignity, rights, and contributions of all individuals, regardless of characteristics including race, ethnicity, caste, color, age, physical characteristics, neurodiversity, disability, sex or gender, gender identity or expression, sexual orientation, language, philosophy or religion, national or social origin, socio-economic position, level of education, or other status. The same privileges of participation are extended to everyone who participates in good faith and in accordance with this Covenant.
+
+
+## Encouraged Behaviors
+
+While acknowledging differences in social norms, we all strive to meet our community's expectations for positive behavior. We also understand that our words and actions may be interpreted differently than we intend based on culture, background, or native language.
+
+With these considerations in mind, we agree to behave mindfully toward each other and act in ways that center our shared values, including:
+
+1. Respecting the purpose of our community, our activities, and our ways of gathering.
+2. Engaging kindly and honestly with others.
+3. Respecting different viewpoints and experiences.
+4. Taking responsibility for our actions and contributions.
+5. Gracefully giving and accepting constructive feedback.
+6. Committing to repairing harm when it occurs.
+7. Behaving in other ways that promote and sustain the well-being of our community.
+
+
+## Restricted Behaviors
+
+We agree to restrict the following behaviors in our community. Instances, threats, and promotion of these behaviors are violations of this Code of Conduct.
+
+1. Harassment. Violating explicitly expressed boundaries or engaging in unnecessary personal attention after any clear request to stop.
+2. Character attacks. Making insulting, demeaning, or pejorative comments directed at a community member or group of people.
+3. Stereotyping or discrimination. Characterizing anyone’s personality or behavior on the basis of immutable identities or traits.
+4. Sexualization. Behaving in a way that would generally be considered inappropriately intimate in the context or purpose of the community.
+5. Violating confidentiality. Sharing or acting on someone's personal or private information without their permission.
+6. Endangerment. Causing, encouraging, or threatening violence or other harm toward any person or group.
+7. Behaving in other ways that threaten the well-being of our community.
+
+### Other Restrictions
+
+1. Misleading identity. Impersonating someone else for any reason, or pretending to be someone else to evade enforcement actions.
+2. Failing to credit sources. Not properly crediting the sources of content you contribute.
+3. Promotional materials. Sharing marketing or other commercial content in a way that is outside the norms of the community.
+4. Irresponsible communication. Failing to responsibly present content which includes, links or describes any other restricted behaviors.
+
+
+## Reporting an Issue
+
+Tensions can occur between community members even when they are trying their best to collaborate. Not every conflict represents a code of conduct violation, and this Code of Conduct reinforces encouraged behaviors and norms that can help avoid conflicts and minimize harm.
+
+When an incident does occur, it is important to report it promptly. To report a possible violation, address it directly in the issues, use the [contact information](https://tumftm.github.io/PointCloudCrafter/#contact) provided in the repository or [contact the institution](https://www.mos.ed.tum.de/en/ftm/contact/) responsible for this repository.
+
+Maintainers and the responsible institute (hereafter referred to as "community moderators") take reports of violations seriously and will make every effort to respond in a timely manner. They will investigate all reports of code of conduct violations. Community moderators will keep investigation and enforcement actions as transparent as possible while prioritizing safety and confidentiality. In order to honor these values, enforcement actions are carried out in private with the involved parties, but communicating to the whole community may be part of a mutually agreed upon resolution.
+
+
+## Addressing and Repairing Harm
+
[…]
