### 2026-05-11 ff7c4c9
add: missing reference to cloudcompare
