+++ docs/paper.bib
@@ -148,3 +148,10 @@ issn={2184-495X},
   isbn={978-989-758-745-0},
   issn={2184-495X},
 }
+
+@misc{CloudCompare2026,
+  author  = {{Girardeau-Montaut, Daniel}},
+  title   = {{CloudCompare} (version 2.13) [GPL software]},
+  year    = {2026},
+  url     = {http://www.cloudcompare.org}
+}
