### 2026-05-11 6914cfc
add: missing reference to cloudcompare
