+++ docs/paper.bib
@@ -150,8 +150,8 @@ issn={2184-495X},
 }
 
 @misc{CloudCompare2026,
-  author  = {{Girardeau-Montaut, Daniel}},
-  title   = {{CloudCompare} (version 2.13) [GPL software]},
+  author  = {Girardeau-Montaut, Daniel},
+  title   = {CloudCompare (version 2.13) [GPL software]},
   year    = {2026},
   url     = {http://www.cloudcompare.org}
 }
