### 2026-05-11 b99d221
add: contribution guidelines
