+++ CONTRIBUTING.md
@@ -1,0 +1,66 @@
+# Contributing <!-- omit from toc -->
+
+Thank you for your interest in contributing! We support all types and sizes of contributions.
+
+As a contributor, here are the guidelines that we would like you to follow:
+
+- [Code of Conduct](#code-of-conduct)
+- [What should I know before I get started?](#what-should-i-know-before-i-get-started)
+- [How can I contribute?](#how-can-i-contribute)
+  - [Bug reports](#bug-reports)
+  - [Pull requests](#pull-requests)
+- [Attribution](#attribution)
+
+Like the tool itself, these guidelines are being actively developed and suggestions for improvement are always welcome!
+
+## Code of Conduct
+
+To ensure the community stays open and inclusive, please follow the [Code of Conduct](CODE_OF_CONDUCT.md).
+
+If you believe that someone in the community has violated the Code of Conduct, please make a report to the listed contacts.
+
+## What should I know before I get started?
+
+If you are new to open source projects, we recommend reading GitHub's [How to Contribute to Open Source guide](https://opensource.guide/how-to-contribute) for an overview of why people contribute to open source projects, what it means to contribute and much more besides.
+
+## How can I contribute?
+
+### Bug reports
+
+Before you report a bug, please search the issue tracker. It is possible that someone has already reported the same issue and that workarounds exist.
+
+When reporting a bug, you should provide a minimal set of instructions to reproduce the issue. Doing so allows us to quickly confirm and focus on the right problem.
+
+If you want to fix the bug by yourself that will be appreciated, but you should discuss possible approaches with the maintainers in the issue before submitting a pull request.
+
+[Creating an issue is straightforward](https://docs.github.com/en/issues/tracking-your-work-with-issues/creating-an-issue#creating-an-issue-from-a-repository).
+
+### Pull requests
+
+You can submit pull requests for small changes such as:
+
+- Minor documentation updates
+- Fixing spelling mistakes
+- Fixing CI failures
+- Fixing warnings detected by compilers or analysis tools
+- Making small changes to a single package
+
+If your pull request is a large change, the following process should be followed:
+
+1. [Create an GitHub issue](https://docs.github.com/en/issues/tracking-your-work-with-issues/creating-an-issue#creating-an-issue-from-a-repository) to propose the change. Doing so allows you to get feedback from the maintainers and to ensure that the proposed change is in line with the design philosophy and current development plans.
+
+2. [Create a pull request](https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/creating-a-pull-request) to implement the changes that references the issue created in step 1.
+
+3. Create documentation for the new addition (if relevant).
+
+Examples of large changes include:
+
+- Adding a new feature
+- Adding a new documentation page or section
+
+Don't forget to also review the required [license](LICENSE).
+
+
+## Attribution
+
+This guideline is adapted from the [Autoware contribution guideline](https://autowarefoundation.github.io/autoware-documentation/main/contributing/).
