### 2026-05-11 121bbd1
add: missing reference to cloudcompare (#76)
* add: missing reference to cloudcompare
