### 2026-05-11 46a1cb3
consistent version numbers + compability indicators
