### 2026-05-11 b4c8f19
add version details to docs
