+++ docs/version_details/index.md
@@ -1,0 +1,20 @@
+# Version Details
+
+This section documents the supported platforms, runtime requirements, and the
+full CLI feature surface of PointCloudCrafter for each release.
+
+The **Latest** page always reflects the current `:latest` Docker image and is
+regenerated on every documentation build. Older version pages are frozen
+snapshots captured at the time of release.
+
+## Available versions
+
+| Version  | Status   | Page                                |
+|----------|----------|-------------------------------------|
+| latest   | rolling  | [Latest](latest.md)                 |
+| v1.2.4 | released | [v1.2.4](v1.2.4.md) |
+| v1.3.4 | released | [v1.3.4](v1.3.4.md) |
+| v1.4.1 | released | [v1.4.1](v1.4.1.md) |
+<!-- VERSION_TABLE_INSERT -->
+
+> New entries are added automatically when a release is published.
