+++ docs/version_details/snippets/help_rosbag_latest.txt
@@ -1,0 +1,2 @@
+(Generated by `make helpgen`. Falls back to this placeholder if the docs build
+cannot reach the Docker image.)
