+++ docs/version_details/v1.2.4.md
@@ -1,0 +1,26 @@
+# v1.2.4
+
+Frozen snapshot captured from `ghcr.io/tumftm/pointcloudcrafter:v1.2.4` at release time.
+
+## Supported platforms
+
+| Component        | Requirement                                              |
+|------------------|----------------------------------------------------------|
+| Operating system | Ubuntu 22.04 (Jammy), x86_64       |
+| glibc            | 2.35                                                 |
+| ROS 2 distro     | Humble                                           |
+| Python           | CPython 3.10.12                                    |
+| Wheel tag        | `cp310-cp310-manylinux_*_x86_64`              |
+| Architectures    | `x86_64` only                                         |
+
+## `pointcloudcrafter-file`
+
+```text
+--8<-- "version_details/snippets/help_file_v1.2.4.txt"
+```
+
+## `pointcloudcrafter-rosbag`
+
+```text
+--8<-- "version_details/snippets/help_rosbag_v1.2.4.txt"
+```
