+++ mkdocs.yml
@@ -22,7 +22,8 @@ markdown_extensions:
     - pymdownx.highlight:
         anchor_linenums: true
     - pymdownx.inlinehilite
-    - pymdownx.snippets
+    - pymdownx.snippets:
+        base_path: ["docs"]
     - pymdownx.superfences
     - pymdownx.emoji:
         emoji_index: !!python/name:material.extensions.emoji.twemoji
@@ -39,6 +40,7 @@ nav:
     - pip_package.md
     - getting_started.md
     - usage.md
+    - Version Details: version_details/index.md
     - third-party.md
     - API: docs_cpp/index.html
 

[…]
