+++ docs/version_details/snippets/help_rosbag_latest.txt
@@ -1,2 +1,63 @@
-(Generated by `make helpgen`. Falls back to this placeholder if the docs build
-cannot reach the Docker image.)
+Create and modify point clouds from rosbags
+Usage: pointcloudcrafter-rosbag [OPTIONS] bag-path out-dir topic-names...
+or
+Usage: ros2 run pointcloudcrafter rosbag [OPTIONS] bag-path out-dir topic-names...
+
+Positionals:
+  bag-path TEXT REQUIRED      Path to ROS 2 bag
+  out-dir TEXT REQUIRED       Output directory for point cloud file(s)
+  topic-names TEXT ... REQUIRED
+                              PointCloud2 topic names
+
+Options:
+  -h,--help                   Print this help message and exit
+
+
+Transforms:
+  -t,--target-frame TEXT      Target TF frame for all point clouds
+  --urdf,--urdf-file TEXT     URDF file providing additional static TF transforms
+  --split FLOAT               Split point cloud into grid cells of given size
+  --tf,--transform-file TEXT  TXT file with transform ([frame_id] r1 r2 r3 x r4 r5 r6 y r7 r8 r9 z)
+
+
+Output:
+  --timestamps                Save point cloud timestamps to a text file
+  --sequential-name           Use sequential file names
+
+
+General:
+  -m,--max-frames INT         Maximum number of frames (-1 = unlimited)
+  -j,--skip-frames INT        Number of frames to skip at the beginning
+  -s,--stride-frames INT      Write every Nth frame
+
+
+File Format:
+  --save-pcd                  Save PCD files (default)
+  --save-ply                  Save PLY files
+  --save-txt                  Save TXT ASCII files
+  --save-kitti                Save KITTI binary files
+  --save-nuscenes             Save nuScenes binary files
+
+
+Filtering:
+  --cb,--crop-box FLOAT FLOAT FLOAT FLOAT FLOAT FLOAT x 6
+                              Crop box [xmin ymin zmin xmax ymax zmax]
+  --cs,--crop-sphere FLOAT    Crop to sphere with given radius
+  --cc,--crop-cylinder FLOAT  Crop to cylinder with given radius
+  --inverse-crop              Inverse crop filters
+  --vf,--voxel-filter FLOAT FLOAT FLOAT x 3
+                              Voxel size [x y z]
+  --orf,--outlier-radius-filter FLOAT INT
+                              Radius outlier removal [radius min_neighbors]
+  --osf,--outlier-stat-filter FLOAT INT
+                              Statistical outlier removal [threshold mean_k]
+
+
+Pip package install example:
+  pointcloudcrafter-rosbag /datasets/bag.mcap /datasets/out/ /points_raw 
+    --voxel-filter 0.1 0.1 0.1 --stride-frames 5
+
+ROS2 install example:
+  ros2 run pointcloudcrafter rosbag /datasets/bag.mcap /datasets/out/ /points_raw 
+    --voxel-filter 0.1 0.1 0.1 --stride-frames 5
+
