### 2026-05-11 45a4b03
move from external repo to release asset for test data
Signed-off-by: Maximilian Leitenstern <maxi.leitenstern@tum.de>
