### 2026-05-11 16fc792
consistent version numbers + compability indicators (#78)
* consistent version numbers + compability indicators

* add version details to docs

* fix navigation in docs
