+++ mkdocs.yml
@@ -22,7 +22,8 @@ markdown_extensions:
     - pymdownx.highlight:
         anchor_linenums: true
     - pymdownx.inlinehilite
-    - pymdownx.snippets
+    - pymdownx.snippets:
+        base_path: ["docs"]
     - pymdownx.superfences
     - pymdownx.emoji:
         emoji_index: !!python/name:material.extensions.emoji.twemoji
@@ -39,6 +40,14 @@ nav:
     - pip_package.md
     - getting_started.md
     - usage.md
+    - Version Details:
+        - Overview: version_details/index.md
+        - Latest: version_details/latest.md
+        # VERSION_NAV_START
+        - v1.4.1: version_details/v1.4.1.md
+        - v1.3.4: version_details/v1.3.4.md
+        - v1.2.4: version_details/v1.2.4.md
+        # VERSION_NAV_END
     - third-party.md
     - API: docs_cpp/index.html
 
