+++ pyproject.toml
@@ -11,7 +11,7 @@ authors = [
     { name = "Dominik Kulmer", email = "dominik.kulmer@tum.de" },
     { name = "Maximilian Leitenstern", email = "maxi.leitenstern@tum.de" },
 ]
-requires-python = ">=3.10"
+requires-python = ">=3.12"
 keywords = [
     "point cloud",
     "LiDAR",
@@ -27,6 +27,9 @@ classifiers = [
     "License :: OSI Approved :: Apache Software License",
     "Operating System :: POSIX :: Linux",
     "Programming Language :: C++",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3 :: Only",
 ]
 dependencies = []
 
