### 2026-05-11 daab471
move from external repo to release asset for test data (#80)
Signed-off-by: Maximilian Leitenstern <maxi.leitenstern@tum.de>
