+++ paper.bib
@@ -1,0 +1,157 @@
+@INPROCEEDINGS{Rusu2011,
+  author={Rusu, Radu Bogdan and Cousins, Steve},
+  booktitle={IEEE International Conference on Robotics and Automation (ICRA)}, 
+  title={3D is here: Point Cloud Library (PCL)}, 
+  year={2011},
+  volume={},
+  number={},
+  pages={1-4},
+  doi={10.1109/ICRA.2011.5980567}
+}
+
+@article{Zhou2018,
+  author    = {Zhou, Qian-Yi and Park, Jaesik and Koltun, Vladlen},
+  title     = {{Open3D}: {A} Modern Library for {3D} Data Processing},
+  journal   = {arXiv:1801.09847},
+  year      = {2018},
+}
+
+@article{Lamprecht2019,
+   author = {Lamprecht, Sebastian},
+   doi = {10.21105/joss.00990},
+   issn = {2475-9066},
+   issue = {36},
+   journal = {Journal of Open Source Software},
+   month = {4},
+   pages = {990},
+   title = {Pyoints: A Python package for point cloud, voxel and raster processing.},
+   volume = {4},
+   year = {2019}
+}
+
+@article{Parigger2024,
+  title={CavitOmiX Drug Discovery: Engineering Antivirals with Enhanced Spectrum and Reduced Side Effects for Arboviral Diseases},
+  author={Parigger, Lena and Krassnigg, Andreas and Hetmann, Michael and Hofmann, Anna and Gruber, Karl and Steinkellner, Georg and Gruber, Christian C},
+  journal={Viruses},
+  volume={16},
+  number={8},
+  pages={1186},
+  year={2024},
+  publisher={MDPI},
+  doi={10.3390/v16081186}
+}
+
+@dataset{nuScenes2019,
+  title={nuScenes: A multimodal dataset for autonomous driving},
+  author={Caesar, Holger and Bankiti, Varun and Lang, Alex and Vora, Sourabh and Liong, Venice and Xu, Qiang and Krishnan, Anush and Pan, Yu and Baldan, Giancarlo and Beijbom, Oscar},
+  booktitle={Proceedings of the IEEE/CVF conference on computer vision and pattern recognition (CVPR)},
+  pages={11621--11631},
+  year={2020}
+}
+
+@inproceedings{KITTI2012,
+  author = {Geiger, Andreas and Lenz, Philip and Urtasun, Raquel},
+  title = {Are we ready for Autonomous Driving? The KITTI Vision Benchmark Suite},
+  booktitle = {Conference on Computer Vision and Pattern Recognition (CVPR)},
+  year = {2012}
+} 
+
+@misc{PDAL2022,
+  author = {PDAL Contributors},
+  title = {PDAL Point Data Abstraction Library},
+  year = 2022,
+  doi = {10.5281/zenodo.2616780},
+  url = {https://doi.org/10.5281/zenodo.2616780}
+}
+
+@article{Pomerleau2013,
+	author = {Pomerleau, Fran{\c c}ois and Colas, Francis and Siegwart, Roland and Magnenat, St{\'e}phane},
+	title = {{Comparing ICP Variants on Real-World Data Sets}},
+	journal = {Autonomous Robots},
+	year = {2013},
+	volume = {34},
+	number = {3},
+	pages = {133--148},
+	month = feb
+}
+
+@article{Schmid2025,
+   author = {Manuel F. Schmid and Jeffrey D. Massey and Marco G. Giometto},
+   doi = {10.21105/joss.07952},
+   issn = {2475-9066},
+   issue = {111},
+   journal = {Journal of Open Source Software},
+   month = {7},
+   pages = {7952},
[…]
