+++ docs_cpp/search/all_15.js
@@ -1,0 +1,6 @@
+var searchData=
+[
+  ['wrapped_5ftype_0',['wrapped_type',['../structCLI_1_1detail_1_1wrapped__type.html',1,'CLI::detail']]],
+  ['wrapped_5ftype_3c_20t_2c_20def_2c_20typename_20std_3a_3aenable_5fif_3c_20is_5fwrapper_3c_20t_20_3e_3a_3avalue_20_3e_3a_3atype_20_3e_1',['wrapped_type&lt; T, def, typename std::enable_if&lt; is_wrapper&lt; T &gt;::value &gt;::type &gt;',['../structCLI_1_1detail_1_1wrapped__type_3_01T_00_01def_00_01typename_01std_1_1enable__if_3_01is__wr7fdd281434af22fc0a3e380b9a3bffc8.html',1,'CLI::detail']]],
+  ['writegridmetadata_2',['writeGridMetadata',['../classpointcloudmodifierlib_1_1Modifier.html#ad9843d19274175a61c37cc4209c60866',1,'pointcloudmodifierlib::Modifier']]]
+];
