+++ docs_cpp/search/searchdata.js
@@ -1,0 +1,39 @@
+var indexSectionsWithContent =
+{
+  0: "_abcdefghijlmnopqrstvw~",
+  1: "abcefhimnoprstvw",
+  2: "c",
+  3: "_abcdefghijlmnopqrstvw~",
+  4: "acdefghilmnoprstv",
+  5: "o",
+  6: "cprv",
+  7: "ev",
+  8: "p"
+};
+
+var indexSectionNames =
+{
+  0: "all",
+  1: "classes",
+  2: "namespaces",
+  3: "functions",
+  4: "variables",
+  5: "enums",
+  6: "enumvalues",
+  7: "groups",
+  8: "pages"
+};
+
+var indexSectionLabels =
+{
+  0: "All",
+  1: "Classes",
+  2: "Namespaces",
+  3: "Functions",
+  4: "Variables",
+  5: "Enumerations",
+  6: "Enumerator",
+  7: "Modules",
+  8: "Pages"
+};
+
