### 2026-05-18 ebfeebf
add ror of TUM to affiliation in paper
Signed-off-by: Maximilian Leitenstern <maxi.leitenstern@tum.de>
