### 2026-05-23 4198617
re-write of research impact statement (#79)
