+++ docs/paper.md
@@ -63,9 +63,12 @@ A comprehensive overview of the available commands, parameters, and usage exampl
 
 # Research impact statement
 
-PointCloudCrafter has been developed to support reproducible preprocessing of large-scale point cloud datasets in research environments. Early versions of the toolkit have been used in student research projects, scientific publications [@Leitenstern2025; @Kulmer2025-1], and the creation of publicly available datasets [@Kulmer2025-2].
-It has been used in a range of internal research workflows related to autonomous driving and robotic perception. These applications include analyzing sensor drivers by inspecting timestamp consistency and packet loss, which manifest as incomplete or irregular point clouds, as well as fusing multiple ROS2 LiDAR topics into a unified point cloud representation using a user-specified static reference frame. The toolkit has been further applied to the creation of custom datasets for evaluating state-of-the-art open-source algorithms in pseudo-labeling, object detection, localization, and mapping. In addition, PointCloudCrafter has been integrated into automated data processing pipelines that convert logged ROS2 data into established dataset formats, such as KITTI, to enable systematic evaluation and labeling for subsequent machine learning training. The software has also been used to generate large-scale, high-resolution point cloud maps by aggregating temporally indexed scans using externally estimated sensor poses.
+PointCloudCrafter was developed to support the reproducible preprocessing of large-scale point cloud datasets in research environments. The software evolved organically within the Institute of Automotive Technology at the Technical University of Munich to replace highly fragmented processing scripts. Driven by the practical requirements of over a dozen student theses and ongoing research projects, the toolkit's feature set was continuously refined through direct user feedback. This iterative, collaborative development established it as a standard framework for point cloud tasks, from raw data extraction to targeted data alteration.
 
+PointCloudCrafter has been applied across diverse robotics domains to solve specific engineering problems. For example, researchers used it to create an automated pipeline that converts ROS2 bag recordings into nuScenes- or KITTI-style datasets. These datasets are used to train and evaluate machine learning perception algorithms [@Tempfli2025] and to create a public dataset for a wide range of applications [@Kulmer2025-2]. In localization research, the tool was used to extract and compile point cloud datasets for LiDAR odometry benchmarking [@Sauerbeck2023]. In mapping, the tool was used for benchmarking and for creating and post-processing large-scale point cloud maps [@Leitenstern2025; @Kulmer2025-1]. Additionally, PointCloudCrafter facilitated the development of novel LiDAR calibration approaches by extracting and isolating specific point cloud objects from continuous ROS2 recordings [@Tahiraj2024; @Tahiraj2026].
+Beyond these projects, PointCloudCrafter is used for analyses such as inspecting sensor drivers for timestamp anomalies or packet losses that result in incomplete or irregular point clouds.
+
+Although it is difficult to track the exact adoption of open-source utilities, PointCloudCrafter has been rigorously tested across domains such as calibration, localization, mapping, and perception.
 By providing a middleware-aware, schema-agnostic CLI toolkit, PointCloudCrafter lowers the barrier for converting raw sensor data into analysis-ready datasets. Its focus on ease of use, reproducibility, and preservation of per-point metadata makes it applicable across multiple domains that rely on three-dimensional point cloud data.
 
 # Availability
