+++ docs/paper.bib
@@ -155,3 +155,38 @@ issn={2184-495X},
   year    = {2026},
   url     = {http://www.cloudcompare.org}
 }
+
+@INPROCEEDINGS{Sauerbeck2023,
+  author={Sauerbeck, Florian and Kulmer, Dominik and Pielmeier, Markus and Leitenstern, Maximilian and Weiß, Christoph and Betz, Johannes},
+  booktitle={2023 IEEE SENSORS}, 
+  title={Multi-LiDAR Localization and Mapping Pipeline for Urban Autonomous Driving}, 
+  year={2023},
+  volume={},
+  number={},
+  pages={1-4},
+  keywords={Location awareness;Laser radar;Pipelines;Urban areas;Software algorithms;Semantics;Real-time systems;Autonomous Vehicles;LiDAR;Maps;Sensor Fusion;SLAM},
+  doi={10.1109/SENSORS56945.2023.10325207}}
+
+@article{Tempfli2025,
+  title={VESPA: Towards un(Human)supervised Open-World Pointcloud Labeling for Autonomous Driving},
+  author={Tempfli, Levente and Rivera, Esteban and Lienkamp, Markus},
+  journal={arXiv preprint arXiv:2507.20397},
+  year={2025}
+}
+
+@INPROCEEDINGS{Tahiraj2024,
+  author={Tahiraj, Ilir and Fent, Felix and Hafemann, Philipp and Ye, Egon and Lienkamp, Markus},
+  booktitle={2024 IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)}, 
+  title={GMMCalib: Extrinsic Calibration of LiDAR Sensors using GMM-based Joint Registration}, 
+  year={2024},
+  pages={8555-8562},
+  doi={10.1109/IROS58592.2024.10801262}}
+
+
+@inproceedings{Tahiraj2026,
+  author={Tahiraj, Ilir and Edinger, Markus and Kulmer, Dominik and Lienkamp, Markus},
+  title={CaLiV: LiDAR-to-Vehicle Calibration of Arbitrary Sensor Setups},
+  booktitle={Proceedings of the 2026 IEEE Intelligent Vehicles Symposium (IV)},
+  year={2026},
+  note={Accepted for publication},
+}
