### 2026-06-02 f1dc1bd
use LiDAR consistently
