### 2026-06-02 edc7e3c
use parallel adjectives
