### 2026-06-02 eb56702
use american spelling for "analyze"
