+++ docs/paper.md
@@ -57,7 +57,7 @@ Internally, all point cloud data is represented using the `PCLPointCloud2` data 
 PointCloudCrafter provides two primary execution modes, *rosbag* and *file*, for streaming and static data workflows, respectively.
 The *rosbag* executable enables extraction of `sensor_msgs::msg::PointCloud2` topics from ROS2 bag recordings. This mode integrates with the `TF2` library [@Foote2013] to perform time-resolved pose lookups based on the recorded transform tree. Each point cloud can be transformed from the sensor coordinate frame into a user-specified fixed reference frame using the pose corresponding to the exact acquisition timestamp. This enables motion-consistent export of point cloud data from dynamic sensor platforms and a timestamp-based fusion of different point cloud sensor topics.
 The *file* executable focuses on batch-oriented processing of static point cloud files. Supported formats include common text and binary formats such as PLY, PCD, OBJ, and TXT, as well as dataset-specific binary formats used in KITTI [@KITTI2012] and nuScenes [@nuScenes2019].
-Both executables support a common set of processing operations implemented in the shared internal codebase. It allows users to apply rigid-body transformations, merge multiple point clouds into a single representation, analyse timestamps, and perform filtering operations. Filters include geometric cropping using box, spherical, or cylindrical regions, voxel grid downsampling, and statistical or radius-based outlier removal. All processing steps are fully parameterized via command-line options, enabling deterministic and reproducible execution across datasets. Due to the schema-agnostic internal representation, per-point attributes, such as sensor-specific metadata or domain-specific properties, are preserved throughout processing whenever supported by the input and output formats.
+Both executables support a common set of processing operations implemented in the shared internal codebase. It allows users to apply rigid-body transformations, merge multiple point clouds into a single representation, analyze timestamps, and perform filtering operations. Filters include geometric cropping using box, spherical, or cylindrical regions, voxel grid downsampling, and statistical or radius-based outlier removal. All processing steps are fully parameterized via command-line options, enabling deterministic and reproducible execution across datasets. Due to the schema-agnostic internal representation, per-point attributes, such as sensor-specific metadata or domain-specific properties, are preserved throughout processing whenever supported by the input and output formats.
 
 A comprehensive overview of the available commands, parameters, and usage examples is provided in the online documentation at \url{https://tumftm.github.io/PointCloudCrafter/usage.html}.
 
