### 2026-06-02 f20bca6
change "it" to "these operations"
