### 2026-06-02 86a1ab9
use monospace for x86_64 and manylinux, math for >=
