### 2026-06-02 1b1b16a
Unified citation style
