+++ docs/paper.bib
@@ -1,89 +1,93 @@
-@INPROCEEDINGS{Rusu2011,
-  author={Rusu, Radu Bogdan and Cousins, Steve},
-  booktitle={IEEE International Conference on Robotics and Automation (ICRA)}, 
-  title={3D is here: Point Cloud Library (PCL)}, 
-  year={2011},
-  volume={},
-  number={},
-  pages={1-4},
-  doi={10.1109/ICRA.2011.5980567}
+@inproceedings{Rusu2011,
+  author = {Rusu, Radu Bogdan and Cousins, Steve},
+  booktitle = {2011 IEEE International Conference on Robotics and Automation},
+  title = {3D Is Here: Point Cloud Library (PCL)},
+  year = {2011},
+  pages = {1-4},
+  doi = {10.1109/ICRA.2011.5980567}
 }
 
-@article{Zhou2018,
-  author    = {Zhou, Qian-Yi and Park, Jaesik and Koltun, Vladlen},
-  title     = {{Open3D}: {A} Modern Library for {3D} Data Processing},
-  journal   = {arXiv:1801.09847},
-  year      = {2018},
+@preprint{Zhou2018,
+  author = {Zhou, Qian-Yi and Park, Jaesik and Koltun, Vladlen},
+  title = {{Open3D}: A Modern Library for {3D} Data Processing},
+  year = {2018},
+  eprint = {1801.09847},
+  eprinttype = {arXiv},
+  eprintclass = {cs.CV}
 }
 
 @article{Lamprecht2019,
    author = {Lamprecht, Sebastian},
    doi = {10.21105/joss.00990},
-   issn = {2475-9066},
-   issue = {36},
+   publisher = {The Open Journal},
+   number = {36},
    journal = {Journal of Open Source Software},
-   month = {4},
    pages = {990},
-   title = {Pyoints: A Python package for point cloud, voxel and raster processing.},
+   title = {Pyoints: A Python Package for Point Cloud, Voxel and Raster Processing},
    volume = {4},
    year = {2019}
 }
 
 @article{Parigger2024,
-  title={CavitOmiX Drug Discovery: Engineering Antivirals with Enhanced Spectrum and Reduced Side Effects for Arboviral Diseases},
-  author={Parigger, Lena and Krassnigg, Andreas and Hetmann, Michael and Hofmann, Anna and Gruber, Karl and Steinkellner, Georg and Gruber, Christian C},
-  journal={Viruses},
-  volume={16},
-  number={8},
-  pages={1186},
-  year={2024},
-  publisher={MDPI},
-  doi={10.3390/v16081186}
+  title = {CavitOmiX Drug Discovery: Engineering Antivirals with Enhanced Spectrum and Reduced Side Effects for Arboviral Diseases},
+  author = {Parigger, Lena and Krassnigg, Andreas and Hetmann, Michael and Hofmann, Anna and Gruber, Karl and Steinkellner, Georg and Gruber, Christian C},
+  journal = {Viruses},
+  volume = {16},
+  number = {8},
+  pages = {1186},
+  year = {2024},
+  publisher = {MDPI},
+  doi = {10.3390/v16081186},
+  url = {https://www.mdpi.com/1999-4915/16/8/1186}
 }
 
-@dataset{nuScenes2019,
-  title={nuScenes: A multimodal dataset for autonomous driving},
-  author={Caesar, Holger and Bankiti, Varun and Lang, Alex and Vora, Sourabh and Liong, Venice and Xu, Qiang and Krishnan, Anush and Pan, Yu and Baldan, Giancarlo and Beijbom, Oscar},
[…]
