+++ docs/paper.bib
@@ -1,7 +1,7 @@
 @inproceedings{Rusu2011,
   author = {Rusu, Radu Bogdan and Cousins, Steve},
   booktitle = {2011 IEEE International Conference on Robotics and Automation},
-  title = {3D Is Here: Point Cloud Library (PCL)},
+  title = {{3D Is Here: Point Cloud Library (PCL)}},
   year = {2011},
   pages = {1-4},
   doi = {10.1109/ICRA.2011.5980567}
@@ -9,7 +9,7 @@
 
 @preprint{Zhou2018,
   author = {Zhou, Qian-Yi and Park, Jaesik and Koltun, Vladlen},
-  title = {{Open3D}: A Modern Library for {3D} Data Processing},
+  title = {{Open3D: A Modern Library for 3D Data Processing}},
   year = {2018},
   eprint = {1801.09847},
   eprinttype = {arXiv},
@@ -23,13 +23,13 @@
    number = {36},
    journal = {Journal of Open Source Software},
    pages = {990},
-   title = {Pyoints: A Python Package for Point Cloud, Voxel and Raster Processing},
+   title = {{Pyoints: A Python Package for Point Cloud, Voxel and Raster Processing}},
    volume = {4},
    year = {2019}
 }
 
 @article{Parigger2024,
-  title = {CavitOmiX Drug Discovery: Engineering Antivirals with Enhanced Spectrum and Reduced Side Effects for Arboviral Diseases},
+  title = {{CavitOmiX Drug Discovery: Engineering Antivirals with Enhanced Spectrum and Reduced Side Effects for Arboviral Diseases}},
   author = {Parigger, Lena and Krassnigg, Andreas and Hetmann, Michael and Hofmann, Anna and Gruber, Karl and Steinkellner, Georg and Gruber, Christian C},
   journal = {Viruses},
   volume = {16},
@@ -42,7 +42,7 @@
 }
 
 @inproceedings{nuScenes2019,
-  title = {nuScenes: A Multimodal Dataset for Autonomous Driving},
+  title = {{nuScenes: A Multimodal Dataset for Autonomous Driving}},
   author = {Caesar, Holger and Bankiti, Varun and Lang, Alex and Vora, Sourabh and Liong, Venice and Xu, Qiang and Krishnan, Anush and Pan, Yu and Baldan, Giancarlo and Beijbom, Oscar},
   booktitle = {2020 IEEE/CVF Conference on Computer Vision and Pattern Recognition (CVPR)},
   pages = {11618-11628},
@@ -52,7 +52,7 @@
 
 @inproceedings{KITTI2012,
   author = {Geiger, Andreas and Lenz, Philip and Urtasun, Raquel},
-  title = {Are We Ready for Autonomous Driving? The KITTI Vision Benchmark Suite},
+  title = {{Are We Ready for Autonomous Driving? The KITTI Vision Benchmark Suite}},
   booktitle = {2012 IEEE Conference on Computer Vision and Pattern Recognition},
   pages = {3354-3361},
   year = {2012},
@@ -60,8 +60,8 @@
 }
 
 @software{PDAL2022,
-  author = {PDAL Contributors},
-  title = {PDAL Point Data Abstraction Library},
+  author = {PDAL{-}Contributors},
+  title = {{PDAL Point Data Abstraction Library}},
   year = 2024,
   version = {2.8.3},
   doi = {10.5281/zenodo.2616780},
@@ -71,7 +71,7 @@
 
 @article{Pomerleau2013,
 	author = {Pomerleau, Fran{\c c}ois and Colas, Francis and Siegwart, Roland and Magnenat, St{\'e}phane},
-	title = {Comparing {ICP} Variants on Real-World Data Sets},
[…]
