### 2026-06-02 942a784
Unified citation style (#86)
* Unified citation style

* title case bib
