### 2026-06-03 0772a29
Suggested edits for the JOSS paper (#87)
* remove hard line breaks

* use real ampersand in bibliography

* change "Especially" to "In particular"

* use LiDAR consistently

* remove hyphens

* use parallel adjectives

* use american spelling for "analyze"

* change "it" to "these operations"

* add "image" after "Docker"

* change to "pull requests"

* improve flow in sentence

* use monospace for x86_64 and manylinux, math for >=

* remove period in date

* add missing "the"
