+++ docs/paper.bib
@@ -87,7 +87,7 @@
    number = {111},
    journal = {Journal of Open Source Software},
    pages = {7952},
-   title = {{PointClouds.jl: Fast \& Flexible Processing of Lidar Data}},
+   title = {{PointClouds.jl: Fast & Flexible Processing of Lidar Data}},
    volume = {10},
    year = {2025}
 }
