### 2026-06-03 6553c25
unified build & testing workflow
