+++ .github/workflows/pypi.yml
@@ -25,7 +25,7 @@ jobs:
             python_tag: "cp312"
     steps:
       - name: Checkout Repo
-        uses: actions/checkout@v5
+        uses: actions/checkout@v6
         with:
           submodules: recursive
 
@@ -98,7 +98,7 @@ jobs:
             ubuntu_version: "24.04"
 
     steps:
-      - uses: actions/checkout@v5
+      - uses: actions/checkout@v6
 
       - name: Download test data from release asset
         run: |
@@ -107,7 +107,7 @@ jobs:
               | tar xz -C test_data
 
       - name: Download wheel
-        uses: actions/download-artifact@v5
+        uses: actions/download-artifact@v8
         with:
           name: pypi-wheel-${{ matrix.python_tag }}
           path: dist/auditwheel
@@ -310,7 +310,7 @@ jobs:
 
     steps:
       - name: Download wheel artifact
-        uses: actions/download-artifact@v5
+        uses: actions/download-artifact@v8
         with:
           pattern: pypi-wheel-*
           path: dist/auditwheel
