+++ .github/workflows/paper-pdf.yml
@@ -10,7 +10,7 @@ jobs:
     name: Paper Draft
     steps:
       - name: Checkout
-        uses: actions/checkout@v4
+        uses: actions/checkout@v6
       - name: Build draft PDF
         uses: openjournals/openjournals-draft-action@master
         with:
@@ -18,7 +18,7 @@ jobs:
           # This should be the path to the paper within your repo.
           paper-path: docs/paper.md
       - name: Upload
-        uses: actions/upload-artifact@v4
+        uses: actions/upload-artifact@v7
         with:
           name: paper
           # This is the output path where Pandoc will write the compiled
