+++ .github/workflows/pypi.yml
@@ -31,7 +31,7 @@ jobs:
 
       - name: Login to Container Registry
         if: github.event_name != 'pull_request'
-        uses: docker/login-action@v3
+        uses: docker/login-action@v4
         with:
           registry: ghcr.io
           username: ${{github.actor}}
@@ -79,7 +79,7 @@ jobs:
             "
 
       - name: Upload wheel for testing
-        uses: actions/upload-artifact@v5
+        uses: actions/upload-artifact@v7
         with:
           name: pypi-wheel-${{ matrix.python_tag }}
           path: dist/auditwheel/*.whl
