### 2026-06-03 71765a8
unified build & testing workflow (#89)
* unified build & testing workflow

* workaround for github api rate limit

* update all actions in all workflows to the latest releases
