+++ .github/workflows/testpypi.yml
@@ -19,10 +19,10 @@ jobs:
             ubuntu_test: "24.04"
 
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v6
 
       - name: Login to GHCR
-        uses: docker/login-action@v3
+        uses: docker/login-action@v4
         with:
           registry: ghcr.io
           username: ${{ github.actor }}
@@ -57,7 +57,7 @@ jobs:
             "
 
       - name: Upload wheel for testing
-        uses: actions/upload-artifact@v4
+        uses: actions/upload-artifact@v7
         with:
           name: testpypi-wheel-${{ matrix.python_tag }}
           path: dist/auditwheel/*.whl
@@ -76,7 +76,7 @@ jobs:
             ubuntu_version: "24.04"
 
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v6
 
       - name: Download test data from release asset
         run: |
@@ -85,7 +85,7 @@ jobs:
               | tar xz -C test_data
 
       - name: Download wheel
-        uses: actions/download-artifact@v4
+        uses: actions/download-artifact@v8
         with:
           name: testpypi-wheel-${{ matrix.python_tag }}
           path: dist/auditwheel
@@ -287,7 +287,7 @@ jobs:
 
     steps:
       - name: Download wheel artifact
-        uses: actions/download-artifact@v4
+        uses: actions/download-artifact@v8
         with:
           pattern: testpypi-wheel-* 
           path: dist/auditwheel
