### 2026-06-08 b2b0e8d
bump version
Signed-off-by: Maximilian Leitenstern <maxi.leitenstern@tum.de>
