### 2026-06-08 ec513b3
Joss release (#90)
* bump version

Signed-off-by: Maximilian Leitenstern <maxi.leitenstern@tum.de>

* fix action orders for docs

Signed-off-by: Maximilian Leitenstern <maxi.leitenstern@tum.de>

---------

Signed-off-by: Maximilian Leitenstern <maxi.leitenstern@tum.de>
