### 2026-06-08 922b5dc
docs: add Version Details for v1.4.3
