+++ docs/version_details/v1.4.3.md
@@ -1,0 +1,26 @@
+# v1.4.3
+
+Frozen snapshot captured from `ghcr.io/tumftm/pointcloudcrafter:v1.4.3` at release time.
+
+## Supported platforms
+
+| Component        | Requirement                                              |
+|------------------|----------------------------------------------------------|
+| Operating system | Ubuntu 24.04 (Noble), x86_64       |
+| glibc            | 2.39                                                 |
+| ROS 2 distro     | Jazzy                                           |
+| Python           | CPython 3.12.3                                    |
+| Wheel tag        | `cp312-cp312-manylinux_*_x86_64`              |
+| Architectures    | `x86_64` only                                         |
+
+## `pointcloudcrafter-rosbag`
+
+```text
+--8<-- "version_details/snippets/help_rosbag_v1.4.3.txt"
+```
+
+## `pointcloudcrafter-file`
+
+```text
+--8<-- "version_details/snippets/help_file_v1.4.3.txt"
+```
