### 2026-06-08 74426af
docs: add Version Details for v1.4.3
