+++ paper.bib
@@ -1,0 +1,193 @@
+@inproceedings{Rusu2011,
+  author = {Rusu, Radu Bogdan and Cousins, Steve},
+  booktitle = {2011 IEEE International Conference on Robotics and Automation},
+  title = {{3D Is Here: Point Cloud Library (PCL)}},
+  year = {2011},
+  pages = {1-4},
+  doi = {10.1109/ICRA.2011.5980567}
+}
+
+@preprint{Zhou2018,
+  author = {Zhou, Qian-Yi and Park, Jaesik and Koltun, Vladlen},
+  title = {{Open3D: A Modern Library for 3D Data Processing}},
+  year = {2018},
+  eprint = {1801.09847},
+  eprinttype = {arXiv},
+  eprintclass = {cs.CV}
+}
+
+@article{Lamprecht2019,
+   author = {Lamprecht, Sebastian},
+   doi = {10.21105/joss.00990},
+   publisher = {The Open Journal},
+   number = {36},
+   journal = {Journal of Open Source Software},
+   pages = {990},
+   title = {{Pyoints: A Python Package for Point Cloud, Voxel and Raster Processing}},
+   volume = {4},
+   year = {2019}
+}
+
+@article{Parigger2024,
+  title = {{CavitOmiX Drug Discovery: Engineering Antivirals with Enhanced Spectrum and Reduced Side Effects for Arboviral Diseases}},
+  author = {Parigger, Lena and Krassnigg, Andreas and Hetmann, Michael and Hofmann, Anna and Gruber, Karl and Steinkellner, Georg and Gruber, Christian C},
+  journal = {Viruses},
+  volume = {16},
+  number = {8},
+  pages = {1186},
+  year = {2024},
+  publisher = {MDPI},
+  doi = {10.3390/v16081186},
+  url = {https://www.mdpi.com/1999-4915/16/8/1186}
+}
+
+@inproceedings{nuScenes2019,
+  title = {{nuScenes: A Multimodal Dataset for Autonomous Driving}},
+  author = {Caesar, Holger and Bankiti, Varun and Lang, Alex and Vora, Sourabh and Liong, Venice and Xu, Qiang and Krishnan, Anush and Pan, Yu and Baldan, Giancarlo and Beijbom, Oscar},
+  booktitle = {2020 IEEE/CVF Conference on Computer Vision and Pattern Recognition (CVPR)},
+  pages = {11618-11628},
+  year = {2020},
+  doi = {10.1109/CVPR42600.2020.01164}
+}
+
+@inproceedings{KITTI2012,
+  author = {Geiger, Andreas and Lenz, Philip and Urtasun, Raquel},
+  title = {{Are We Ready for Autonomous Driving? The KITTI Vision Benchmark Suite}},
+  booktitle = {2012 IEEE Conference on Computer Vision and Pattern Recognition},
+  pages = {3354-3361},
+  year = {2012},
+  doi = {10.1109/CVPR.2012.6248074}
+}
+
+@software{PDAL2022,
+  author = {PDAL{-}Contributors},
+  title = {{PDAL Point Data Abstraction Library}},
+  year = 2024,
+  version = {2.8.3},
+  doi = {10.5281/zenodo.2616780},
+  publisher = {Zenodo},
+  url = {https://doi.org/10.5281/zenodo.2616780}
+}
+
+@article{Pomerleau2013,
+	author = {Pomerleau, Fran{\c c}ois and Colas, Francis and Siegwart, Roland and Magnenat, St{\'e}phane},
+	title = {{Comparing ICP Variants on Real-World Data Sets}},
+	journal = {Autonomous Robots},
+	year = {2013},
+	volume = {34},
+	number = {3},
[…]
