+++ docs_cpp/search/namespaces_0.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['cli_3a_3aenums_0',['enums',['../namespaceCLI_1_1enums.html',1,'CLI']]]
+];
