+++ docs_cpp/search/functions_d.js
@@ -1,0 +1,7 @@
+var searchData=
+[
+  ['name_0',['name',['../classCLI_1_1Validator.html#a8fc0ddaaa61f5dc5a7fa527d697991da',1,'CLI::Validator::name(std::string validator_name)'],['../classCLI_1_1Validator.html#acb5ba8d1f332769ed85efa7ff6d90218',1,'CLI::Validator::name(std::string validator_name) const'],['../classCLI_1_1App.html#a9a5972fee0abc892c43919c8d4e53a6d',1,'CLI::App::name()']]],
+  ['needs_1',['needs',['../classCLI_1_1Option.html#ad81f0537b30d16f02a066473ef3ad5dc',1,'CLI::Option::needs(Option *opt)'],['../classCLI_1_1Option.html#aec2946d8cf9744bc55a3b8954df09999',1,'CLI::Option::needs(std::string opt_name)'],['../classCLI_1_1Option.html#ad9fb912d3f5fcf2b882cbb220412a3be',1,'CLI::Option::needs(A opt, B opt1, ARG... args)']]],
+  ['non_5fmodifying_2',['non_modifying',['../classCLI_1_1Validator.html#a356d979d984dffff3489caf4baca9160',1,'CLI::Validator']]],
+  ['nonpositional_3',['nonpositional',['../classCLI_1_1Option.html#a6208cfc282d4029741775092a4e9c52d',1,'CLI::Option']]]
+];
