### 2026-06-25 311e360
modify ROS 2 refs, ghcr link and bibtex entries
