+++ docs/paper.bib
@@ -13,7 +13,8 @@
   year = {2018},
   eprint = {1801.09847},
   eprinttype = {arXiv},
-  eprintclass = {cs.CV}
+  eprintclass = {cs.CV},
+  doi = {10.48550/arXiv.1801.09847}
 }
 
 @article{Lamprecht2019,
@@ -171,7 +172,8 @@
   year = {2025},
   eprint = {2507.20397},
   eprinttype = {arXiv},
-  eprintclass = {cs.CV}
+  eprintclass = {cs.CV},
+  doi = {10.48550/arXiv.2507.20397}
 }
 
 @inproceedings{Tahiraj2024,
@@ -189,5 +191,6 @@
   eprinttype = {arXiv},
   eprint = {2504.01987},
   eprintclass = {cs.RO},
-  year = {2026}
+  year = {2026},
+  doi = {10.48550/arXiv.2504.01987}
 }
