### 2026-06-26 1cdcd47
adjust status badge
Signed-off-by: Maximilian Leitenstern <maxi.leitenstern@tum.de>
