### 2026-06-26 15ebd18
Joss release (#92)
add citation file
replace joss review badge with actual DOI
