+++ docs_cpp/search/pages_0.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['pointcloudcrafter_0',['PointCloudCrafter',['../index.html',1,'']]]
+];
