+++ docs_cpp/search/variables_8.js
@@ -1,0 +1,5 @@
+var searchData=
+[
+  ['labels_5f_0',['labels_',['../classCLI_1_1FormatterBase.html#a2ee79421fa339b227d6cb9ca1bce4ad1',1,'CLI::FormatterBase']]],
+  ['lnames_5f_1',['lnames_',['../classCLI_1_1Option.html#ae06feaf7ca62a5497514a09f44724430',1,'CLI::Option']]]
+];

[…]
